requirements.txt
anthropic>=0.40.0
langgraph>=0.2.0
pydantic>=2.0.0
openpyxl>=3.1.0
python-pptx>=0.6.21
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.6
